refactor: move OffersSection into TasksPage
and fix tests.

OffersSection is initially rendered on the TaskCard.

This makes it unnessecary to be lazy loaded
